requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --extra=dev --output-file=requirements/dev.txt --resolver=backtracking pyproject.toml
#
asgiref==3.5.2
    # via django
asttokens==2.1.0
    # via stack-data
attrs==22.1.0
    # via pytest
backcall==0.2.0
    # via ipython
black==22.10.0
    # via santa-unchained (pyproject.toml)
click==8.1.3
    # via black
coverage[toml]==6.5.0
    # via pytest-cov
decorator==5.1.1
    # via ipython
django==4.1.3
    # via
    #   django-admin-display
    #   django-debug-toolbar
    #   django-extensions
    #   santa-unchained (pyproject.toml)
django-admin-display==1.3.0
    # via santa-unchained (pyproject.toml)
django-debug-toolbar==3.7.0
    # via santa-unchained (pyproject.toml)
django-extensions==3.2.1
    # via santa-unchained (pyproject.toml)
environs==9.5.0
    # via santa-unchained (pyproject.toml)
exceptiongroup==1.0.4
    # via pytest
execnet==1.9.0
    # via pytest-xdist
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via
    #   pytest-factoryboy
    #   santa-unchained (pyproject.toml)
faker==15.3.3
    # via factory-boy
flake8==5.0.4
    # via
    #   flake8-formatter-abspath
    #   santa-unchained (pyproject.toml)
flake8-formatter-abspath==1.0.1
    # via santa-unchained (pyproject.toml)
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipython==8.6.0
    # via santa-unchained (pyproject.toml)
isort==5.10.1
    # via santa-unchained (pyproject.toml)
jedi==0.18.1
    # via ipython
marshmallow==3.19.0
    # via environs
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   marshmallow
    #   pytest
    #   pytest-sugar
parso==0.8.3
    # via jedi
pathspec==0.10.2
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.4
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.32
    # via ipython
psycopg2-binary==2.9.5
    # via santa-unchained (pyproject.toml)
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-sugar
    #   pytest-xdist
    #   santa-unchained (pyproject.toml)
pytest-cov==4.0.0
    # via santa-unchained (pyproject.toml)
pytest-django==4.5.2
    # via santa-unchained (pyproject.toml)
pytest-factoryboy==2.5.0
    # via santa-unchained (pyproject.toml)
pytest-sugar==0.9.6
    # via santa-unchained (pyproject.toml)
pytest-xdist==3.0.2
    # via santa-unchained (pyproject.toml)
python-dateutil==2.8.2
    # via faker
python-dotenv==0.21.0
    # via environs
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.1
    # via ipython
termcolor==2.1.0
    # via pytest-sugar
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.4.0
    # via pytest-factoryboy
wcwidth==0.2.5
    # via prompt-toolkit
